+++ smpp/smpptest/server.go
@@ -1,4 +1,4 @@
-package main
+package smpptest
 
 import (
 	"bufio"
